Move files to moe_training/nvfp4_training
