+++ test/prototype/moe_training/nvfp4_training/__init__.py
@@ -1,0 +1,1 @@
+"""NVFP4 training prototype tests."""
